First time C64 Datasette works. A simple Hello World program has been loaded and run.
The datasette implementation has been rebuilt for all type of computer. Now the paralell datasettes don't exist anymore and the the standard implementation has been rebuilt.

VICII Sound works.

+++ src/CORE/global.cpp
@@ -317,7 +317,7 @@ std::string MCHEmul::tableFormat (const MCHEmul::Strings& s, const std::string& 
 	std::string result;
 
 	MCHEmul::Strings cs = s;
-	if (ne != -1 && cs.size () > ne)
+	if (ne != -1 && cs.size () > (size_t) ne)
 		cs = MCHEmul::Strings (cs.end () - ne, cs.end ());
 
 	for (size_t i = 0; i < cs.size (); i += sb)
